qcom wdog: delay wdog bark time interval
System will sometime crash due to the qcom watchdog, when the
system overload is very busy, it's easy to not response dog pet
action and cause the system reboot.

So we change watchdog bark time to 16 sec to reduce dog bit chance
when the system is a little busy.

Change-Id: Id6d076cc069c2aeb00290ccfbbde862863a23922
Reviewed-on: https://gerrit.mot.com/2876598
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Reviewed-by: Guobin Zhang <zhanggb@motorola.com>
Tested-by: Jira Key
Submit-Approved: Jira Key